added mapped materials to inspector panel

+++ src/app/types.hpp
@@ -5,7 +5,7 @@
 #ifndef OPENGLRENDERINGENGINE_TYPES_HPP
 #define OPENGLRENDERINGENGINE_TYPES_HPP
 
-using index_t = size_t;
+using index_t = uint32_t;
 using uuid64_t = uint64_t;
 using gpu_tex_handle64_t = uint64_t;
